Unify documents.permissions and shipments.permissions method in DRY optic
