+++ apps/documents/permissions.py
@@ -1,15 +1,28 @@
+"""
+Copyright 2019 ShipChain, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
 import logging
-from django.conf import settings
-from rest_framework import permissions, status
+from rest_framework import permissions
 
 from apps.shipments.models import Shipment
-from apps.authentication import get_jwt_from_request
-from apps.permissions import has_owner_access
+
+from apps.permissions import check_has_shipment_owner_access
 
 
 LOG = logging.getLogger('transmission')
-
-PROFILES_URL = f'{settings.PROFILES_URL}/api/v1/wallet'
 
 
 class UserHasPermission(permissions.BasePermission):
@@ -20,9 +33,7 @@ class UserHasPermission(permissions.BasePermission):
     def has_object_permission(self, request, view, obj):
         shipment = obj.shipment
 
-        return has_owner_access(request, obj) or has_owner_access(request, shipment) or \
-            self.is_shipper(request, shipment) or self.is_carrier(request, shipment) or \
-            self.is_moderator(request, shipment)
+        return check_has_shipment_owner_access(request, shipment)
 
     def has_permission(self, request, view):
 
@@ -36,33 +47,4 @@ class UserHasPermission(permissions.BasePermission):
             LOG.warning(f'User: {request.user}, is trying to access documents of not found shipment: {shipment_id}')
             return False
 
-        return has_owner_access(request, shipment) or self.is_shipper(request, shipment) or \
-            self.is_carrier(request, shipment) or self.is_moderator(request, shipment)
-
-    def is_carrier(self, request, shipment):
-        """
-        Custom permission for carrier documents management access
-        """
-        response = settings.REQUESTS_SESSION.get(f'{PROFILES_URL}/{shipment.carrier_wallet_id}/?is_active',
-                                                 headers={'Authorization': f'JWT {get_jwt_from_request(request)}'})
-
-        return response.status_code == status.HTTP_200_OK
-
-    def is_moderator(self, request, shipment):
-        """
-        Custom permission for moderator documents management access
-        """
-        if shipment.moderator_wallet_id:
-            response = settings.REQUESTS_SESSION.get(f'{PROFILES_URL}/{shipment.moderator_wallet_id}/?is_active',
-                                                     headers={'Authorization': f'JWT {get_jwt_from_request(request)}'})
-
-        return shipment.moderator_wallet_id and response.status_code == status.HTTP_200_OK
-
-    def is_shipper(self, request, shipment):
-        """
-        Custom permission for shipper documents management access
-        """
-        response = settings.REQUESTS_SESSION.get(f'{PROFILES_URL}/{shipment.shipper_wallet_id}/?is_active',
-                                                 headers={'Authorization': f'JWT {get_jwt_from_request(request)}'})
-
-        return response.status_code == status.HTTP_200_OK
+        return check_has_shipment_owner_access(request, shipment)
